rename web3 => web3registry
